+++ src/ConfDirective.cc
@@ -700,20 +700,28 @@ static const struct BackupTimeDirective: InheritableDirective {
     const size_t dash = cc.bits[1].find('-');
     if(dash == std::string::npos)
       throw SyntaxError("expected EARLIEST-LATEST");
-    
+
     int earliest = parseTimeOfDay(cc.bits[1].substr(0, dash));
-    int latest = parseTimeOfDay(cc.bits[1].substr(dash+1, std::string::npos));
+    int latest = parseTimeOfDay(cc.bits[1].substr(dash + 1, std::string::npos));
 
     if(latest == 0)
       latest = 86400;
     if(earliest > latest)
       throw SyntaxError("earliest backup time (" + formatTimeOfDay(earliest)
-                        + ") is later than latest backup time (" +formatTimeOfDay(latest)
-                        + ")");
+                        + ") is later than latest backup time ("
+                        + formatTimeOfDay(latest) + ")");
     cc.context->earliest = earliest;
     cc.context->latest = latest;
   }
 } backup_time_directive;
+
+/** @brief The @c group directive */
+static const struct GroupDirective: public InheritableDirective {
+  GroupDirective(): InheritableDirective("group", 1, 1) {}
+  void set(ConfContext &cc) const override {
+    cc.context->group = cc.bits[1];
+  }
+} group_directive;
 
 // Host directives ------------------------------------------------------------
 
@@ -730,14 +738,6 @@ static const struct HostDirective: public ConfDirective {
     cc.host->hostname = cc.bits[1];
   }
 } host_directive;
-
-/** @brief The @c group directive */
-static const struct GroupDirective: public HostOnlyDirective {
-  GroupDirective(): HostOnlyDirective("group", 1, 1) {}
-  void set(ConfContext &cc) const override {
-    cc.host->group = cc.bits[1];
-  }
-} group_directive;
 
 /** @brief The @c hostname directive */
 static const struct HostnameDirective: public HostOnlyDirective {